FAQ_Chat_Bot/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
twilio==8.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
